requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
pydantic
python-dotenv
psycopg2-binary